Fix all 25 outstanding bug work items
Fixes across the codebase:
- UB: grid index overflow, attr ID truncation, BMP signed shifts,
  perf timing underflow, glyph cache empty cluster overflow
- Memory: FT_Face lifetime guard, SDL dialog use-after-free and path
  leak, Vulkan resource leaks (sampled image, sync objects), RPC late
  response leak, megacity callbacks lifetime guard
- PTY: EINTR handling for poll/write, shutdown fd close ordering
- Correctness: empty APPDATA config path, CLI arg crash on bad input,
  megacity image size, empty mesh Vulkan violation, scrollback attr
  compaction, scrollback stale stride, mpack reserve OOM cap, PPI
  float equality, megacity renderer config data race

All work items moved to plans/work-items-complete/.

+++ libs/draxul-types/src/bmp.cpp
@@ -45,7 +45,8 @@ bool write_bmp_rgba(const std::filesystem::path& path, const CapturedFrame& fram
 
     std::filesystem::create_directories(path.parent_path());
 
-    const auto image_size = static_cast<uint32_t>(frame.width * frame.height * 4);
+    const auto image_size = static_cast<uint32_t>(
+        static_cast<uint64_t>(frame.width) * frame.height * 4);
     const auto file_size = kBmpHeaderSize + kBmpInfoSize + image_size;
 
     std::vector<uint8_t> bytes;
@@ -97,7 +98,10 @@ std::optional<CapturedFrame> read_bmp_rgba(const std::filesystem::path& path)
         return static_cast<uint16_t>(bytes[offset] | (bytes[offset + 1] << 8));
     };
     auto read_u32 = [&](size_t offset) -> uint32_t {
-        return static_cast<uint32_t>(bytes[offset] | (bytes[offset + 1] << 8) | (bytes[offset + 2] << 16) | (bytes[offset + 3] << 24));
+        return static_cast<uint32_t>(bytes[offset])
+            | (static_cast<uint32_t>(bytes[offset + 1]) << 8)
+            | (static_cast<uint32_t>(bytes[offset + 2]) << 16)
+            | (static_cast<uint32_t>(bytes[offset + 3]) << 24);
     };
     auto read_i32 = [&](size_t offset) -> int32_t {
         return static_cast<int32_t>(read_u32(offset));
